move util function so they are shareable with typescript
