+++ src/wasm-lib/kcl/src/std/utils.rs
@@ -442,3 +442,25 @@ mod tests {
         assert_eq!(angle_end.degrees().round(), 180.0);
     }
 }
+
+pub type Coords2d = [f64; 2];
+
+pub fn is_points_ccw_wasm(points: &[f64]) -> i32 {
+    // CCW is positive as that the Math convention
+    // assert!(points.len() % 2 == 0, "Points array should have even length");
+    // TODO the same function is defined in `src/wasm-lib/kcl/src/std/sketch.rs`, but I'm too bad a rust
+    // to figure out how to import one into the other
+
+    let mut sum = 0.0;
+    for i in 0..(points.len() / 2) {
+        let point1 = [points[2 * i], points[2 * i + 1]];
+        let point2 = [points[(2 * i + 2) % points.len()], points[(2 * i + 3) % points.len()]];
+        sum += (point2[0] + point1[0]) * (point2[1] - point1[1]);
+    }
+    sum.signum() as i32
+}
+
+pub fn is_points_ccw(points: &[Coords2d]) -> i32 {
+    let flattened_points: Vec<f64> = points.iter().flat_map(|&p| vec![p[0], p[1]]).collect();
+    is_points_ccw_wasm(&flattened_points)
+}
